+++ utilities/ImpactUtilities.groovy
@@ -16,6 +16,7 @@ import com.ibm.dbb.dependency.internal.*
 @Field def buildUtils= loadScript(new File("BuildUtilities.groovy"))
 @Field def metadataUtils= loadScript(new File("MetadatastoreUtilities.groovy"))
 @Field def dependencyScannerUtils= loadScript(new File("DependencyScannerUtilities.groovy"))
+@Field def matcherUtils= loadScript(new File("MatcherUtilities.groovy"))
 @Field String hashPrefix = ':githash:'
 
 
@@ -79,7 +80,7 @@ def createImpactBuildList() {
 				if (props.verbose) println "** Performing impact analysis on changed file $changedFile"
 
 				// get exclude list
-				List<PathMatcher> excludeMatchers = buildUtils.createPathMatcherPattern(props.excludeFileList)
+				List<PathMatcher> excludeMatchers = matcherUtils.createPathMatcherPattern(props.excludeFileList)
 
 				// list of impacts
 				String impactSearch = props.getFileProperty('impactSearch', changedFile)
@@ -92,7 +93,7 @@ def createImpactBuildList() {
 					// only add impacted files that have a build script mapped to it
 					if (ScriptMappings.getScriptName(impactFile)) {
 						// only add impacted files, that are in scope of the build.
-						if (!buildUtils.matches(impactFile, excludeMatchers)){
+						if (!matcherUtils.matches(impactFile, excludeMatchers)){
 
 							// calculate abbreviated gitHash for impactFile
 							filePattern = FileSystems.getDefault().getPath(impactFile).getParent().toString()
@@ -140,7 +141,7 @@ def createImpactBuildList() {
 
 
 					// get excludeListe
-					List<PathMatcher> excludeMatchers = buildUtils.createPathMatcherPattern(props.excludeFileList)
+					List<PathMatcher> excludeMatchers = matcherUtils.createPathMatcherPattern(props.excludeFileList)
 
 					logicalFileList.each { logicalFile ->
 						def impactFile = logicalFile.getFile()
@@ -148,7 +149,7 @@ def createImpactBuildList() {
 						// only add impacted files that have a build script mapped to it
 						if (ScriptMappings.getScriptName(impactFile)) {
 							// only add impacted files, that are in scope of the build.
-							if (!buildUtils.matches(impactFile, excludeMatchers)){
+							if (!matcherUtils.matches(impactFile, excludeMatchers)){
 								buildSet.add(impactFile)
 								if (props.verbose) println "** $impactFile is impacted by changed property $changedProp. Adding to build list."
 							}
@@ -415,13 +416,13 @@ def calculateChangedFiles(BuildResult lastBuildResult, boolean calculateConcurre
 		def mode = null
 
 		// make sure file is not an excluded file
-		List<PathMatcher> excludeMatchers = buildUtils.createPathMatcherPattern(props.excludeFileList)
+		List<PathMatcher> excludeMatchers = matcherUtils.createPathMatcherPattern(props.excludeFileList)
 
 		if (props.verbose) println "*** Changed files for directory $dir $msg:"
 		changed.each { file ->
 			(file, mode) = fixGitDiffPath(file, dir, true, null)
 			if ( file != null ) {
-				if ( !buildUtils.matches(file, excludeMatchers)) {
+				if ( !matcherUtils.matches(file, excludeMatchers)) {
 					changedFiles << file
 					if (!calculateConcurrentChanges) githashBuildableFilesMap.addFilePattern(abbrevCurrent, file)
 					if (props.verbose) println "**** $file"
@@ -438,7 +439,7 @@ def calculateChangedFiles(BuildResult lastBuildResult, boolean calculateConcurre
 
 		if (props.verbose) println "*** Deleted files for directory $dir $msg:"
 		deleted.each { file ->
-			if ( !buildUtils.matches(file, excludeMatchers)) {
+			if ( !matcherUtils.matches(file, excludeMatchers)) {
 				(file, mode) = fixGitDiffPath(file, dir, false, mode)
 				deletedFiles << file
 				if (props.verbose) println "**** $file"
@@ -447,7 +448,7 @@ def calculateChangedFiles(BuildResult lastBuildResult, boolean calculateConcurre
 
 		if (props.verbose) println "*** Renamed files for directory $dir $msg:"
 		renamed.each { file ->
-			if ( !buildUtils.matches(file, excludeMatchers)) {
+			if ( !matcherUtils.matches(file, excludeMatchers)) {
 				(file, mode) = fixGitDiffPath(file, dir, false, mode)
 				renamedFiles << file
 				if (props.verbose) println "**** $file"
@@ -579,8 +580,8 @@ def fixGitDiffPath(String file, String dir, boolean mustExist, mode) {
  */
 def boolean shouldCalculateImpacts(String changedFile){
 	// retrieve Pathmaters from property and check
-	List<PathMatcher> nonImpactingFiles = buildUtils.createPathMatcherPattern(props.skipImpactCalculationList)
-	onskipImpactCalculationList = buildUtils.matches(changedFile, nonImpactingFiles)
+	List<PathMatcher> nonImpactingFiles = matcherUtils.createPathMatcherPattern(props.skipImpactCalculationList)
+	onskipImpactCalculationList = matcherUtils.matches(changedFile, nonImpactingFiles)
 
 	// return false if changedFile found in skipImpactCalculationList
 	if (onskipImpactCalculationList) return false
