+++ utilities/MatcherUtilities.groovy
@@ -1,0 +1,66 @@
+import java.nio.file.FileSystems
+import java.nio.file.Path
+import java.nio.file.PathMatcher
+import java.util.regex.Pattern
+
+/**
+ * createPathMatcherPattern
+ * Generic method to build PathMatcher from a build property
+ */
+
+def createPathMatcherPattern(String property) {
+	List<PathMatcher> pathMatchers = new ArrayList<PathMatcher>()
+	if (property) {
+		property.split(',').each{ filePattern ->
+			if (!filePattern.startsWith('glob:') || !filePattern.startsWith('regex:'))
+				filePattern = "glob:$filePattern"
+			PathMatcher matcher = FileSystems.getDefault().getPathMatcher(filePattern)
+			pathMatchers.add(matcher)
+		}
+	}
+	return pathMatchers
+}
+
+/**
+ * create List of Regex Patterns
+ */
+
+def createMatcherPatterns(String property) {
+	List<Pattern> patterns = new ArrayList<Pattern>()
+	if (property) {
+		property.split(',').each{ patternString ->
+			Pattern pattern = Pattern.compile(patternString);
+			patterns.add(pattern)
+		}
+	}
+	return patterns
+}
+
+/**
+ * matches
+ * Generic method to validate if a file is matching any pathmatchers
+ *
+ */
+def matches(String file, List<PathMatcher> pathMatchers) {
+	def result = pathMatchers.any { matcher ->
+		Path path = FileSystems.getDefault().getPath(file);
+		if ( matcher.matches(path) )
+		{
+			return true
+		}
+	}
+	return result
+}
+
+/**
+ * match a String against a list of patterns
+ */
+ def matchesPattern(String name, List<Pattern> patterns) {
+	def result = patterns.any { pattern ->
+		if (pattern.matcher(name).matches())
+		{
+			return true
+		}
+	}
+	return result
+ }
